LZW
Copied over LZW, hooked it up to encrypt, and simplified the code a
bit. Tested on some big files, so I think it still works. Still not
perfect implementation of LZW, since it doesn’t store frequencies of
prefixes, but it’s completely correct (file sizes may be a bit big
sometimes, but encode and decode are still inverses).